Refactored a bit CardUtils.
This class is actually useless. It's not possible to check things on a card but CardInstance needs to be used instead. This is because a card could be an artifact but some effects transforms it in a creature.

This class is useful only for admin screen of for DeckFactory methods (which possibly should not be in Game module anyway.)